+++ pit_components/lib/components/adv_future_builder.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 typedef void ResetFunction();
 typedef Widget WidgetBuilder(BuildContext context);
-typedef Future<bool> FutureExecutor(BuildContext context, int attempt);
+typedef Future<bool> FutureExecutor(BuildContext context);
 
 class ResetRemote {
   ResetFunction reset;
@@ -22,26 +22,28 @@ class AdvFutureBuilder extends StatefulWidget {
 }
 
 class _AdvFutureBuilderState extends State<AdvFutureBuilder> {
-  int executionAttempt = 1;
+  bool _futureExecuted = false;
 
   @override
   void initState() {
     super.initState();
 
     widget.remote?.reset = () {
-      executionAttempt = 1;
+      _futureExecuted = false;
     };
   }
 
   @override
   Widget build(BuildContext context) {
-    widget.futureExecutor(context, executionAttempt).then((available) {
-      if (available) {
-        setState(() {
-          executionAttempt++;
-        });
-      }
-    });
+    if (!_futureExecuted) {
+      _futureExecuted = true;
+      widget.futureExecutor(context).then((needRedraw) {
+        if (needRedraw) {
+          setState(() {
+          });
+        }
+      });
+    }
 
     return widget.widgetBuilder(context);
   }