+++ pit_components/lib/components/controllers/adv_text_field_controller.dart
@@ -16,7 +16,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   String get hint => value.hint;
@@ -34,7 +35,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   String get label => value.label;
@@ -52,7 +54,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   String get error => value.error;
@@ -70,7 +73,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   int get maxLength => value.maxLength;
@@ -88,7 +92,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   bool get maxLengthEnforced => value.maxLengthEnforced;
@@ -106,7 +111,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   int get maxLines => value.maxLines;
@@ -124,7 +130,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   bool get enable => value.enable;
@@ -142,7 +149,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   TextAlign get alignment => value.alignment;
@@ -160,7 +168,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: newAlignment,
         obscureText: this.obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   bool get obscureText => value.obscureText;
@@ -178,7 +187,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: newObscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   Widget get prefixIcon => value.prefixIcon;
@@ -196,7 +206,8 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: this.obscureText,
         prefixIcon: newPrefixIcon,
-        suffixIcon: this.suffixIcon);
+        suffixIcon: this.suffixIcon,
+        selection: this.selection);
   }
 
   Widget get suffixIcon => value.suffixIcon;
@@ -214,48 +225,73 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
         alignment: this.alignment,
         obscureText: obscureText,
         prefixIcon: this.prefixIcon,
-        suffixIcon: newSuffixIcon);
+        suffixIcon: newSuffixIcon,
+        selection: this.selection);
+  }
+
+  TextSelection get selection => value.selection;
+
+  set selection(TextSelection newSelection) {
+    if (newSelection.start > text.length || newSelection.end > text.length)
+      throw FlutterError('invalid text selection: $newSelection');
+    value = value.copyWith(
+        text: this.text,
+        hint: this.hint,
+        label: this.label,
+        error: this.error,
+        maxLength: this.maxLength,
+        maxLengthEnforced: this.maxLengthEnforced,
+        maxLines: this.maxLines,
+        enable: this.enable,
+        alignment: this.alignment,
+        obscureText: obscureText,
+        prefixIcon: this.prefixIcon,
+        suffixIcon: this.suffixIcon,
+        selection: newSelection);
   }
 
   AdvTextFieldController(
       {String text,
-        String hint,
-        String label,
-        String error,
-        int maxLength,
-        bool maxLengthEnforced,
-        int maxLines,
-        bool enable,
-        TextAlign alignment,
-        bool obscureText,
-        Widget prefixIcon,
-        Widget suffixIcon})
+      String hint,
+      String label,
+      String error,
+      int maxLength,
+      bool maxLengthEnforced,
+      int maxLines,
+      bool enable,
+      TextAlign alignment,
+      bool obscureText,
+      Widget prefixIcon,
+        Widget suffixIcon,
+        TextSelection selection})
       : super(text == null &&
-      hint == null &&
-      label == null &&
-      error == null &&
-      maxLength == null &&
-      maxLengthEnforced == null &&
-      maxLines == null &&
-      enable == null &&
-      alignment == null &&
-      obscureText == null &&
-      prefixIcon == null &&
-      suffixIcon == null
-      ? AdvTextFieldEditingValue.empty
-      : new AdvTextFieldEditingValue(
-      text: text,
-      hint: hint,
-      label: label,
-      error: error,
-      maxLength: maxLength,
-      maxLengthEnforced: maxLengthEnforced ?? false,
-      maxLines: maxLines,
-      enable: enable ?? true,
-      alignment: alignment ?? TextAlign.left,
-      obscureText: obscureText ?? false,
-      prefixIcon: prefixIcon,
-      suffixIcon: suffixIcon));
+                hint == null &&
+                label == null &&
+                error == null &&
+                maxLength == null &&
+                maxLengthEnforced == null &&
+                maxLines == null &&
+                enable == null &&
+                alignment == null &&
+                obscureText == null &&
+                prefixIcon == null &&
+      suffixIcon == null &&
+      selection == null
+            ? AdvTextFieldEditingValue.empty
+            : new AdvTextFieldEditingValue(
+                text: text,
+                hint: hint,
+                label: label,
+                error: error,
+                maxLength: maxLength,
+                maxLengthEnforced: maxLengthEnforced ?? false,
+                maxLines: maxLines,
+                enable: enable ?? true,
+                alignment: alignment ?? TextAlign.left,
+                obscureText: obscureText ?? false,
+                prefixIcon: prefixIcon,
+      suffixIcon: suffixIcon,
+      selection: selection = const TextSelection.collapsed(offset: -1),));
 
   AdvTextFieldController.fromValue(AdvTextFieldEditingValue value)
       : super(value ?? AdvTextFieldEditingValue.empty);
@@ -269,17 +305,18 @@ class AdvTextFieldController extends ValueNotifier<AdvTextFieldEditingValue> {
 class AdvTextFieldEditingValue {
   const AdvTextFieldEditingValue(
       {this.text = '',
-        this.hint = '',
-        this.label = '',
-        this.error = '',
-        this.maxLength,
-        this.maxLengthEnforced = false,
-        this.maxLines,
-        this.enable = true,
-        this.alignment = TextAlign.left,
-        this.obscureText = false,
-        this.prefixIcon,
-        this.suffixIcon});
+      this.hint = '',
+      this.label = '',
+      this.error = '',
+      this.maxLength,
+      this.maxLengthEnforced = false,
+      this.maxLines,
+      this.enable = true,
+      this.alignment = TextAlign.left,
+      this.obscureText = false,
+      this.prefixIcon,
+        this.suffixIcon,
+        this.selection = const TextSelection.collapsed(offset: -1)});
 
   final String text;
   final String hint;
@@ -293,23 +330,25 @@ class AdvTextFieldEditingValue {
   final bool obscureText;
   final Widget prefixIcon;
   final Widget suffixIcon;
+  final TextSelection selection;
 
   static const AdvTextFieldEditingValue empty =
-  const AdvTextFieldEditingValue();
+      const AdvTextFieldEditingValue();
 
   AdvTextFieldEditingValue copyWith(
       {String text,
-        String hint,
-        String label,
-        String error,
-        int maxLength,
-        bool maxLengthEnforced,
-        int maxLines,
-        bool enable,
-        TextAlign alignment,
-        bool obscureText,
-        Widget prefixIcon,
-        Widget suffixIcon}) {
+      String hint,
+      String label,
+      String error,
+      int maxLength,
+      bool maxLengthEnforced,
+      int maxLines,
+      bool enable,
+      TextAlign alignment,
+      bool obscureText,
+      Widget prefixIcon,
+        Widget suffixIcon,
+        TextSelection selection}) {
     return new AdvTextFieldEditingValue(
         text: text ?? this.text,
         hint: hint ?? this.hint,
@@ -322,7 +361,8 @@ class AdvTextFieldEditingValue {
         alignment: alignment ?? this.alignment,
         obscureText: obscureText ?? this.obscureText,
         prefixIcon: prefixIcon ?? this.prefixIcon,
-        suffixIcon: suffixIcon ?? this.suffixIcon);
+        suffixIcon: suffixIcon ?? this.suffixIcon,
+        selection: selection ?? this.selection);
   }
 
   AdvTextFieldEditingValue.fromValue(AdvTextFieldEditingValue copy)
@@ -337,11 +377,12 @@ class AdvTextFieldEditingValue {
         this.alignment = copy.alignment,
         this.obscureText = copy.obscureText,
         this.prefixIcon = copy.prefixIcon,
-        this.suffixIcon = copy.suffixIcon;
+        this.suffixIcon = copy.suffixIcon,
+        this.selection = copy.selection;
 
   @override
   String toString() =>
-      '$runtimeType(text: \u2524$text\u251C, \u2524$hint\u251C, \u2524$label\u251C, \u2524$error\u251C, maxLength: $maxLength, maxLengthEnforced: $maxLengthEnforced, maxLines: $maxLines, enable: $enable, alignment: $alignment, obscureText: $obscureText, prefixIcon: $prefixIcon, suffixIcon: $suffixIcon)';
+      '$runtimeType(text: \u2524$text\u251C, \u2524$hint\u251C, \u2524$label\u251C, \u2524$error\u251C, maxLength: $maxLength, maxLengthEnforced: $maxLengthEnforced, maxLines: $maxLines, enable: $enable, alignment: $alignment, obscureText: $obscureText, prefixIcon: $prefixIcon, suffixIcon: $suffixIcon, selection: $selection)';
 
   @override
   bool operator ==(dynamic other) {
@@ -359,7 +400,8 @@ class AdvTextFieldEditingValue {
         typedOther.alignment == alignment &&
         typedOther.obscureText == obscureText &&
         typedOther.prefixIcon == prefixIcon &&
-        typedOther.suffixIcon == suffixIcon;
+        typedOther.suffixIcon == suffixIcon &&
+        typedOther.selection == selection;
   }
 
   @override
@@ -375,5 +417,6 @@ class AdvTextFieldEditingValue {
       alignment.hashCode,
       obscureText.hashCode,
       prefixIcon.hashCode,
-      suffixIcon.hashCode);
+      suffixIcon.hashCode,
+      selection.hashCode);
 }